Extend the vpc endpoints security group to the EC2 frontend cluster
The addition of the VPC endpoints for the new frontend cluster caused
an issue with the EC2 cluster starting new tasks, since it was trying
and failing to talk to the VPC endpoint for secrets manager.

This commit extends the vpc endpoint security group to allow
connections from the EC2 cluster.