Peer the frontend and backend VPCs
This is in preparation for allowing the new frontend Fargate cluster
to talk to the authentication and logging APIs (via new ECS services)
without having to go over the internet. This means that access can be
granted by security group ID, rather than relying on granting access
via public IP address.

There's three peering connections per environment. The frontend VPC is
peered to the backend VPC in each region, and additionally, the
frontend VPC in the dublin region is paired with the backend VPC in
the london region, to allow FreeRADIUS to talk to the logging API.