Add Sync Certs Job To AWS Codebuild
This commit adds the "sync cert" job that previously existed in Concourse to
AWS codebuild. The codebuild job copies the TLS/SSL certificates from the certs
directory in the private govwifi-build repo (govwifi-build/passwords/certs/)
and adds them to S3 buckets in eu-west-1 and eu-west-2.

The certificates are then copied to the GovWifi radius servers when the servers are restarted.
The codebuild job is created in each GovWifi environment and in each GovWifi
region. For example there will  be a sync-cert codebuild job in:
Production eu-west-1
Production eu-west-2
Staging eu-west-1
Staging eu-west-2

The job utilises the previously existing shell scripts found in govwifi-build/ci/tasks/scripts/