Change all references to authorisation-api to authentication-api
We are making any references to the authentication api infrastructure
consistent. At present we are using the name "authentication-api" and
"authorization-api" to refer to the same things, which is confusing. This
change is part of rectifying that. It updates the name of the cluster, service
and related resources. More information is available in this card:

https://technologyprogramme.atlassian.net/jira/software/projects/GW/boards/251?selectedIssue=GW-204