Reinstate separate staging and production repos for ECR images
In order to ensure that code is not released to production before it is fully
tested in staging we need to have two separate image repos for the GovWifi:
production and staging. This ensures that if the ECS tasks are restarted in
staging before testing is complete, they will not launch with an untested
image.

Prior to this commit the task definitions for staging and production pointed at
the same image repo.