Add Automatic Snapshot Volume Backups for Prometheus and Grafana
What
This enables the data for Prometheus, and the Grafana configuration to be
recovered if the volumes or instances are deleted. The snapshots are created
and managed using Amazon Data Lifecycle Manager, which incurs no extra cost
(https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/snapshot-lifecycle.html).

Why
It is best practice to have a backup of our data