Automatically Add Radius IPs To Secrets
These are used by the smoke tests codebuild job to test our system. This speeds
up our Disaster Recovery process, by giving us one less secret that we have to
manually enter into the Secrets Manager when we create a new account.

The aws.dublin provider has been added as the smoke tests only exist in one
region (London), but need to access a secret created in the Ireland region
(eu-west-1). The addition of this provider to the govwifi-smoketests module
allows this cross regional secret retrieval to happen.