Add Healthchecks For Documentation & Product Pages
We need to be alerted whenever the following sites are down:
https://www.wifi.service.gov.uk/
https://dev-docs.wifi.service.gov.uk/
https://docs.wifi.service.gov.uk/

This PR adds a route53 healthcheck, which is linked to a cloudwatch alarm. The
cloudwatch alarm is in turn linked to our critical alerts sns topic. We will be
notified by both slack and email if these sites go down. Outside of office
hours these alerts are sent to pager duty.